rastertolabel: Add support for Munbyn ITPP941 label printers
The implementation has been tested to generate the same output as the
Munbyn MacOS driver version 1.2.5.

The only tested device is a 203 dpi label printer. Support for 300 dpi
printers (ITPP941P) has been implemented but not tested.

Some Rollo and Beeprt printers might be supported as well, considering
that they all are using the same closed-source Linux filter for CUPS,
called rastertolabelbeeprt in some packages.

The PPD files for those printers set cupsModelNumber to 20. This
implementation is using the same value.